绑角色绑定权限和用户

+++ resources/views/RbacRole/index.blade.php
@@ -22,19 +22,17 @@
                 </div>
             </div>
             <div class="box-body">
-                <div class="table-responsive">
-                    <table class="table table-hover table-striped table-condensed" id="tblRbacRole">
-                        <thead>
-                        <tr>
-                            <th>新建时间</th>
-                            <th>编号</th>
-                            <th>名称</th>
-                            <th></th>
-                        </tr>
-                        </thead>
-                        <tbody></tbody>
-                    </table>
-                </div>
+                <table class="table table-hover table-striped table-condensed" id="tblRbacRole">
+                    <thead>
+                    <tr>
+                        <th>新建时间</th>
+                        <th>编号</th>
+                        <th>名称</th>
+                        <th></th>
+                    </tr>
+                    </thead>
+                    <tbody></tbody>
+                </table>
             </div>
         </div>
     </section>
@@ -120,27 +118,29 @@
          * @param {string} uuid
          */
         function fnDelete(uuid = "") {
-            let loading = layer.msg('处理中……', {time: 0,});
-            $.ajax({
-                url: `{{ url("rbacRole") }}/${uuid}`,
-                type: 'delete',
-                data: {},
-                async: true,
-                success: res => {
-                    console.log(`{{ url("rbacRole") }}/${uuid} success:`, res);
-                    layer.close(loading);
-                    layer.msg(res['msg'], {time: 1000,}, function () {
-                        tblRbacRole.ajax.reload();
-                    });
-                },
-                error: err => {
-                    console.log(`{{ url("rbacRole") }}/${uuid} fail:`, err);
-                    layer.close(loading);
-                    layer.msg(err["responseJSON"], {time: 1500,}, () => {
-                        if (err.status === 401) location.href = '{{ route('web.Authorization:GetLogin') }}';
-                    });
-                },
-            });
+            if (uuid && confirm("删除不可恢复，是否确定？")) {
+                let loading = layer.msg('处理中……', {time: 0,});
+                $.ajax({
+                    url: `{{ url("rbacRole") }}/${uuid}`,
+                    type: 'delete',
+                    data: {},
+                    async: true,
+                    success: res => {
+                        console.log(`{{ url("rbacRole") }}/${uuid} success:`, res);
+                        layer.close(loading);
+                        layer.msg(res['msg'], {time: 1000,}, function () {
+                            tblRbacRole.ajax.reload();
+                        });
+                    },
+                    error: err => {
+                        console.log(`{{ url("rbacRole") }}/${uuid} fail:`, err);
+                        layer.close(loading);
+                        layer.msg(err["responseJSON"], {time: 1500,}, () => {
+                            if (err.status === 401) location.href = '{{ route('web.Authorization:GetLogin') }}';
+                        });
+                    },
+                });
+            }
         }
     </script>
 @endsection
